+++ src/Sekiban.Core/Query/QueryModel/INextAggregateListQueryAsync.cs
@@ -9,6 +9,3 @@ public interface INextAggregateListQueryAsync<TAggregatePayload, TOutput> : INex
     public Task<ResultBox<IEnumerable<TOutput>>> HandleFilterAsync(IEnumerable<AggregateState<TAggregatePayload>> list, IQueryContext context);
     public Task<ResultBox<IEnumerable<TOutput>>> HandleSortAsync(IEnumerable<TOutput> filteredList, IQueryContext context);
 }
-public interface ITenantNextAggregateListQueryAsync<TAggregatePayload, TOutput> : INextAggregateQueryCommon<TAggregatePayload, TOutput>,
-    INextListQueryCommon<TOutput>, INextQueryAsyncGeneral,
-    ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
