+++ src/Sekiban.Core/Query/QueryModel/INextAggregateListQueryWithPagingAsync.cs
@@ -3,5 +3,3 @@ namespace Sekiban.Core.Query.QueryModel;
 
 public interface INextAggregateListQueryWithPagingAsync<TAggregatePayload, TOutput> : INextAggregateListQueryAsync<TAggregatePayload, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
-public interface ITenantNextAggregateListQueryWithPagingAsync<TAggregatePayload, TOutput> : INextAggregateListQueryAsync<TAggregatePayload, TOutput>,
-    IQueryPagingParameterCommon, ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
