+++ src/Sekiban.Core/Query/QueryModel/INextAggregateQuery.cs
@@ -8,5 +8,3 @@ public interface INextAggregateQuery<TAggregatePayload, TOutput> : INextAggregat
     public QueryListType QueryListType => QueryListType.ActiveOnly;
     public ResultBox<TOutput> HandleFilter(IEnumerable<AggregateState<TAggregatePayload>> list, IQueryContext context);
 }
-public interface ITenantNextAggregateQuery<TAggregatePayload, TOutput> : INextAggregateQueryCommon<TAggregatePayload, TOutput>,
-    INextQueryCommon<TOutput>, ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
