+++ src/Sekiban.Core/Query/QueryModel/INextGeneralListQueryWithPaging.cs
@@ -1,5 +1,3 @@
 namespace Sekiban.Core.Query.QueryModel;
 
 public interface INextGeneralListQueryWithPaging<TOutput> : INextGeneralListQuery<TOutput>, IQueryPagingParameterCommon where TOutput : notnull;
-public interface ITenantNextGeneralListQueryWithPaging<TOutput> : INextGeneralListQuery<TOutput>, IQueryPagingParameterCommon, ITenantQueryCommon
-    where TOutput : notnull;
