+++ src/Sekiban.Core/Query/QueryModel/INextGeneralListQueryWithPagingAsync.cs
@@ -2,5 +2,3 @@ namespace Sekiban.Core.Query.QueryModel;
 
 public interface INextGeneralListQueryWithPagingAsync<TOutput> : INextGeneralListQueryAsync<TOutput>, IQueryPagingParameterCommon
     where TOutput : notnull;
-public interface ITenantNextGeneralListQueryWithPagingAsync<TOutput> : INextGeneralListQueryAsync<TOutput>, IQueryPagingParameterCommon,
-    ITenantQueryCommon where TOutput : notnull;
