+++ src/Sekiban.Core/Query/QueryModel/INextMultiProjectionListQueryWithPagingAsync.cs
@@ -5,7 +5,3 @@ public interface
     INextMultiProjectionListQueryWithPagingAsync<TMultiProjectionPayloadCommon, TOutput> :
     INextMultiProjectionListQueryAsync<TMultiProjectionPayloadCommon, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
-public interface
-    ITenantNextMultiProjectionListQueryWithPagingAsync<TMultiProjectionPayloadCommon, TOutput> :
-    INextMultiProjectionListQueryAsync<TMultiProjectionPayloadCommon, TOutput>, IQueryPagingParameterCommon,
-    ITenantQueryCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
