+++ src/Sekiban.Core/Query/QueryModel/INextMultiProjectionListWithPagingQuery.cs
@@ -5,7 +5,3 @@ public interface
     INextMultiProjectionListWithPagingQuery<TMultiProjectionPayloadCommon, TOutput> :
     INextMultiProjectionListQuery<TMultiProjectionPayloadCommon, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
-public interface
-    ITenantNextMultiProjectionListWithPagingQuery<TMultiProjectionPayloadCommon, TOutput> :
-    INextMultiProjectionListQuery<TMultiProjectionPayloadCommon, TOutput>, IQueryPagingParameterCommon,
-    ITenantQueryCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
