+++ src/Sekiban.Core/Query/QueryModel/INextSingleProjectionListQueryWithPagingAsync.cs
@@ -5,7 +5,3 @@ public interface
     INextSingleProjectionListQueryWithPagingAsync<TSingleProjectionPayloadCommon, TOutput> :
     INextSingleProjectionListQueryAsync<TSingleProjectionPayloadCommon, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
-public interface
-    ITenantNextSingleProjectionListQueryWithPagingAsync<TSingleProjectionPayloadCommon, TOutput> :
-    INextSingleProjectionListQueryAsync<TSingleProjectionPayloadCommon, TOutput>, IQueryPagingParameterCommon,
-    ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
