+++ src/Sekiban.Core/Query/QueryModel/INextSingleProjectionListWithPagingQuery.cs
@@ -5,7 +5,3 @@ public interface
     INextSingleProjectionListWithPagingQuery<TSingleProjectionPayloadCommon, TOutput> :
     INextSingleProjectionListQuery<TSingleProjectionPayloadCommon, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
-public interface
-    ITenantNextSingleProjectionListWithPagingQuery<TSingleProjectionPayloadCommon, TOutput> :
-    INextSingleProjectionListQuery<TSingleProjectionPayloadCommon, TOutput>, IQueryPagingParameterCommon,
-    ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
