+++ src/Sekiban.Core/Query/QueryModel/INextSingleProjectionQuery.cs
@@ -9,7 +9,3 @@ public interface
     public QueryListType QueryListType => QueryListType.ActiveOnly;
     public ResultBox<TOutput> HandleFilter(IEnumerable<SingleProjectionState<TSingleProjectionPayloadCommon>> list, IQueryContext context);
 }
-public interface
-    ITenantNextSingleProjectionQuery<TSingleProjectionPayloadCommon, TOutput> : INextSingleProjectionQueryCommon<TSingleProjectionPayloadCommon, TOutput>,
-    INextQueryCommon<TOutput>, ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon
-;
