+++ src/Sekiban.Core/Query/QueryModel/ITenantNextAggregateListQuery.cs
@@ -1,0 +1,5 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextAggregateListQuery<TAggregatePayload, TOutput> : INextAggregateQueryCommon<TAggregatePayload, TOutput>,
+    INextListQueryCommon<TOutput>, ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
