+++ src/Sekiban.Core/Query/QueryModel/ITenantNextAggregateListQueryWithPaging.cs
@@ -1,0 +1,5 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextAggregateListQueryWithPaging<TAggregatePayload, TOutput> : INextAggregateListQuery<TAggregatePayload, TOutput>,
+    IQueryPagingParameterCommon, ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
