+++ src/Sekiban.Core/Query/QueryModel/ITenantNextAggregateListQueryWithPagingAsync.cs
@@ -1,0 +1,5 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextAggregateListQueryWithPagingAsync<TAggregatePayload, TOutput> : INextAggregateListQueryAsync<TAggregatePayload, TOutput>,
+    IQueryPagingParameterCommon, ITenantQueryCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
