+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralListQuery.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralListQuery<TOutput> : INextGeneralQueryCommon<TOutput>, INextListQueryCommon<TOutput>, ITenantQueryCommon
+    where TOutput : notnull;
