+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralListQueryAsync.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralListQueryAsync<TOutput> : INextGeneralQueryCommon<TOutput>, INextListQueryCommon<TOutput>, INextQueryAsyncGeneral,
+    ITenantQueryCommon where TOutput : notnull;
