+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralListQueryWithPaging.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralListQueryWithPaging<TOutput> : INextGeneralListQuery<TOutput>, IQueryPagingParameterCommon, ITenantQueryCommon
+    where TOutput : notnull;
