+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralListQueryWithPagingAsync.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralListQueryWithPagingAsync<TOutput> : INextGeneralListQueryAsync<TOutput>, IQueryPagingParameterCommon,
+    ITenantQueryCommon where TOutput : notnull;
