+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralQuery.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralQuery<TOutput> : INextGeneralQueryCommon<TOutput>, INextQueryCommon<TOutput>, ITenantQueryCommon
+    where TOutput : notnull;
