+++ src/Sekiban.Core/Query/QueryModel/ITenantNextGeneralQueryAsync.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface ITenantNextGeneralQueryAsync<TOutput> : INextGeneralQueryCommon<TOutput>, INextQueryCommon<TOutput>, INextQueryAsyncGeneral,
+    ITenantQueryCommon where TOutput : notnull;
