+++ src/Sekiban.Core/Query/QueryModel/ITenantNextMultiProjectionListQueryAsync.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Query.MultiProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextMultiProjectionListQueryAsync<TMultiProjectionPayloadCommon, TOutput> :
+    INextMultiProjectionQueryCommon<TMultiProjectionPayloadCommon, TOutput>, INextListQueryCommon<TOutput>,
+    ITenantQueryCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
