+++ src/Sekiban.Core/Query/QueryModel/ITenantNextMultiProjectionQuery.cs
@@ -1,0 +1,6 @@
+using Sekiban.Core.Query.MultiProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextMultiProjectionQuery<TMultiProjectionPayloadCommon, TOutput> : INextMultiProjectionQueryCommon<TMultiProjectionPayloadCommon, TOutput>,
+    INextQueryCommon<TOutput>, ITenantQueryCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
