+++ src/Sekiban.Core/Query/QueryModel/ITenantNextMultiProjectionQueryAsync.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Query.MultiProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextMultiProjectionQueryAsync<TMultiProjectionPayloadCommon, TOutput> :
+    INextMultiProjectionQueryCommon<TMultiProjectionPayloadCommon, TOutput>, INextQueryCommon<TOutput>,
+    ITenantQueryCommon where TOutput : notnull where TMultiProjectionPayloadCommon : IMultiProjectionPayloadCommon;
