+++ src/Sekiban.Core/Query/QueryModel/ITenantNextSingleProjectionListQuery.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Query.SingleProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextSingleProjectionListQuery<TSingleProjectionPayloadCommon, TOutput> :
+    INextSingleProjectionQueryCommon<TSingleProjectionPayloadCommon, TOutput>, INextListQueryCommon<TOutput>,
+    ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
