+++ src/Sekiban.Core/Query/QueryModel/ITenantNextSingleProjectionQuery.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Query.SingleProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextSingleProjectionQuery<TSingleProjectionPayloadCommon, TOutput> : INextSingleProjectionQueryCommon<TSingleProjectionPayloadCommon, TOutput>,
+    INextQueryCommon<TOutput>, ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon
+;
