+++ src/Sekiban.Core/Query/QueryModel/ITenantNextSingleProjectionQueryAsync.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Query.SingleProjections;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface
+    ITenantNextSingleProjectionQueryAsync<TSingleProjectionPayloadCommon, TOutput> :
+    INextSingleProjectionQueryCommon<TSingleProjectionPayloadCommon, TOutput>, INextQueryCommon<TOutput>,
+    ITenantQueryCommon where TOutput : notnull where TSingleProjectionPayloadCommon : ISingleProjectionPayloadCommon;
